requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
setuptools==80.9.0
urllib3==2.5.0
wheel==0.45.1